+++ src/App.jsx
@@ -1,17 +1,13 @@
+import TodoForm from './TodoForm';
+import TodoList from './TodoList';
 import './App.css';
 
 function App() {
-  const todos = [
-    {id: 1, title: "review resources"},
-    {id: 2, title: "take notes"},
-    {id: 3, title: "code out app"}
-  ];
   return (
   <div>
     <h1>Todo List</h1>
-    <ul>
-      {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
-    </ul>
+    <TodoForm></TodoForm>
+    <TodoList></TodoList>
   </div>
   )
 }
