+++ src/TodoForm.jsx
@@ -1,0 +1,10 @@
+function TodoForm() {
+    return (
+        <form>
+            <label htmlFor="todoTitle">Todo</label>
+            <input id="todoTitle" type="text" />
+            <button>Add Todo</button>
+        </form>
+    )
+}
+export default TodoForm;
